requirements.txt

pytest
pytest-cov
openai
python-dotenv
